Enable Non-Pixel-Perfect Sprites, Use Float Coordinates, and Add Physics with Heron (#40)
This is a big update that switches to using floating points for sprite coordinates and also allows opting-out of the pixel perfect alignment. This allows you to use pixel-perfect alignment for non-moving objects, but to allow for higher definition movement on moving object such as players and projectiles.

The rendering stuff is pretty much done, but we still have to rework the collision system now. We were going to add a better way to access and manage collisions anyway so we'll throw those changes in this PR as well.